Change from BigInteger to Integer in Users table

+++ database/migrations/2024_06_26_214916_create_mesa_folios_table.php
@@ -19,11 +19,11 @@ return new class extends Migration {
             $table->foreignUuid('libro_digital_id');
             $table->unsignedBigInteger('mesa_id')->nullable();
             $table->foreignUuid('master_materia_id');
-            $table->unsignedBigInteger('presidente_id')->nullable();
-            $table->unsignedBigInteger('vocal_1_id')->nullable();
-            $table->unsignedBigInteger('vocal_2_id')->nullable();
-            $table->unsignedBigInteger('coordinador_id')->nullable();
-            $table->unsignedBigInteger('operador_id')->nullable();
+            $table->unsignedInteger('presidente_id')->nullable();
+            $table->unsignedInteger('vocal_1_id')->nullable();
+            $table->unsignedInteger('vocal_2_id')->nullable();
+            $table->unsignedInteger('coordinador_id')->nullable();
+            $table->unsignedInteger('operador_id')->nullable();
 
             $table->softDeletes();
             $table->timestamps();
